+++ view/tws_administration_view/lib/view/articles/trucks/options/adapters/trucks_article_table_adapter.dart
@@ -690,6 +690,11 @@ final class _TableAdapter extends TWSArticleTableAdapter<Truck> {
                 set.truckCommonNavigation?.situationNavigation?.name ?? '---',
           ),
           TWSPropertyViewer(
+            label: 'Insurance',
+            value:
+                set.insuranceNavigation?.policy ?? '---',
+          ),
+          TWSPropertyViewer(
             label: 'Location',
             value: set.truckCommonNavigation?.locationNavigation?.name ?? '---',
           ),